+++ mangadap/tests/conftest.py
@@ -4,6 +4,7 @@ import pytest
 import os
 import mangadap
 
+# TODO: I don't think this is ever used...
 def pytest_report_header(config):
     drp_root = os.path.join(os.environ['MANGA_SPECTRO_REDUX'], os.environ['MANGADRP_VER'])
     if not os.path.isdir(drp_root):